fix(config): add new config DTO for Actioman
This commit introduces a new configuration DTO (Data Transfer Object) for the Actioman
application. The `Config` DTO defines the structure of the configuration options that
can be used to configure the Actioman server, including settings for the server port,
hostname, SSL, and integrations.

The new configuration DTO is added to the `config-file/dtos/config.dto.ts` file, and
the corresponding export is added to the `package.json` file to make it accessible
to other parts of the application.